+++ demo/src/main/java/com/orange/ods/demo/ui/guidelines/Guideline.kt
@@ -13,19 +13,27 @@ package com.orange.ods.demo.ui.guidelines
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import androidx.compose.runtime.Immutable
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
 import com.orange.ods.demo.R
 import com.orange.ods.demo.ui.MainDestinations
 
 @Immutable
-data class Guideline(
+enum class Guideline(
     @StringRes val titleRes: Int,
     @DrawableRes val imageRes: Int,
     val route: String
-)
+) {
+    Color(R.string.guideline_colors, R.drawable.il_colors, MainDestinations.GUIDELINE_COLORS),
+    Typography(R.string.guideline_typography, R.drawable.il_typography, MainDestinations.GUIDELINE_TYPOGRAPHY);
+    //Imagery(R.string.guideline_imagery, R.drawable.il_imagery),
+    //Iconography(R.string.guideline_iconography, R.drawable.il_iconography);
 
-val guidelines = listOf(
-    Guideline(R.string.guideline_colors, R.drawable.il_colors, MainDestinations.GUIDELINE_COLORS),
-    Guideline(R.string.guideline_typography, R.drawable.il_typography, MainDestinations.GUIDELINE_TYPOGRAPHY),
-    //Guideline(R.string.guideline_imagery, R.drawable.il_imagery),
-    //Guideline(R.string.guideline_iconography, R.drawable.il_iconography)
-)
+    val imageBackgroundColor = Color(0xff1b1b1b)
+
+    val imageContentScale: ContentScale
+        get() = when (this) {
+            Color -> ContentScale.FillBounds
+            Typography -> ContentScale.Fit
+        }
+}